modified some file

+++ threads.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   threads.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ryutaro320515 <ryutaro320515@student.42    +#+  +:+       +#+        */
+/*   By: rmatsuba <rmatsuba@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 10:54:44 by ryutaro3205       #+#    #+#             */
-/*   Updated: 2024/05/09 14:50:55 by ryutaro3205      ###   ########.fr       */
+/*   Updated: 2024/05/09 21:25:57 by rmatsuba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	thread_philo(t_philo *philo, t_env *env)
 {
-	size_t		i;
+	long		i;
 	pthread_t	monitor;
 
 	pthread_create(&monitor, NULL, &monitoring, philo);
